+++ src/main/java/com/example/warehouse/service/impl/UnitServiceImpl.java
@@ -1,5 +1,7 @@
 package com.example.warehouse.service.impl;
 
+import com.example.warehouse.dto.unit.UnitDTO;
+import com.example.warehouse.mapper.UnitMapper;
 import com.example.warehouse.model.Unit;
 import com.example.warehouse.repository.UnitRepository;
 import com.example.warehouse.service.UnitService;
@@ -8,47 +10,60 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class UnitServiceImpl implements UnitService {
 
     final private UnitRepository unitRepository;
+    final private UnitMapper unitMapper;
 
     @Autowired
-    public UnitServiceImpl(UnitRepository unitRepository) {
+    public UnitServiceImpl(UnitRepository unitRepository, UnitMapper unitMapper) {
         this.unitRepository = unitRepository;
+        this.unitMapper = unitMapper;
     }
 
     @Override
-    public Unit createUnit(Unit unit) {
-        return unitRepository.save(unit);
+    public UnitDTO createUnit(UnitDTO unitDTO) {
+        return unitMapper.toUnitDTO(unitRepository.save(unitMapper.toUnit(unitDTO)));
     }
 
     @Override
-    public List<Unit> getAllUnits() {
-        return (List<Unit>) unitRepository.findAll();
+    public List<UnitDTO> getAllUnits() {
+        return unitRepository.findAll().stream()
+                .map(unitMapper::toUnitDTO)
+                .collect(Collectors.toList());
     }
 
     @Override
-    public Optional<Unit> getUnitById(Long id) {
-        return unitRepository.findById(id);
+    public UnitDTO getUnitById(Long id) {
+        return unitMapper.toUnitDTO(unitRepository.findById(id).orElse(null));
     }
 
     @Override
-    public Unit updateUnitById(Long id, Unit newUnit) {
+    public UnitDTO updateUnitById(Long id, UnitDTO unitDTO) {
         Optional<Unit> unit = unitRepository.findById(id);
 
         if (unit.isPresent()) {
-            unit.get().setName(newUnit.getName());
+            unitMapper.updateUnitFromDTO(unitDTO, unit.get());
 
-            return unitRepository.save(unit.get());
+            return unitMapper.toUnitDTO(unitRepository.save(unit.get()));
         }
-        return unitRepository.save(newUnit);
+        return null;
     }
 
     @Override
-    public void deleteUnitById(Long id) {
-        unitRepository.deleteById(id);
+    public boolean deleteUnitById(Long id) {
+        Optional<Unit> unit = unitRepository.findById(id);
+
+        if (unit.isPresent()) {
+            unitRepository.deleteById(id);
+
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
